+++ runtime/prumo_runtime/commands/__init__.py
@@ -1,3 +1,4 @@
+from prumo_runtime.commands.auth_google import run_auth_google
 from prumo_runtime.commands.briefing import run_briefing
 from prumo_runtime.commands.context_dump import run_context_dump
 from prumo_runtime.commands.migrate import run_migrate
@@ -6,6 +7,7 @@ from prumo_runtime.commands.setup import run_setup
 from prumo_runtime.commands.snapshot_refresh import run_snapshot_refresh
 
 __all__ = [
+    "run_auth_google",
     "run_briefing",
     "run_context_dump",
     "run_migrate",